create a PIInsightsService
