+++ tools/PI/DevHome.PI/Pages/InsightsPage.xaml
@@ -20,7 +20,7 @@
 
         <TextBlock x:Uid="InsightsHeaderTextBlock" FontSize="{StaticResource SubtitleTextBlockFontSize}" FontWeight="SemiBold" Margin="0,0,0,8"/>
 
-        <ItemsControl ItemsSource="{x:Bind ViewModel.InsightsList}" Grid.Row="1">
+        <ItemsControl ItemsSource="{x:Bind ViewModel.InsightsService.InsightsList}" Grid.Row="1">
             <ItemsControl.ItemTemplate>
                 <DataTemplate x:DataType="models:Insight">
                     <Expander ExpandDirection="Down" HorizontalAlignment="Stretch" VerticalAlignment="Top" Padding="0,6">
