+++ tools/PI/DevHome.PI/ViewModels/BarWindowViewModel.cs
@@ -1,27 +1,23 @@
 ﻿// Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-using System;
 using System.Collections.ObjectModel;
 using System.Collections.Specialized;
 using System.ComponentModel;
 using System.Diagnostics;
-using System.Globalization;
-using System.Text;
 using CommunityToolkit.Mvvm.ComponentModel;
 using CommunityToolkit.Mvvm.Input;
 using DevHome.Common.Extensions;
 using DevHome.PI.Helpers;
 using DevHome.PI.Models;
+using DevHome.PI.Services;
 using Microsoft.UI.Xaml;
 using Microsoft.UI.Xaml.Controls;
 using Microsoft.UI.Xaml.Media.Imaging;
 using Serilog;
 using Windows.Graphics;
 using Windows.System;
-using Windows.Win32;
 using Windows.Win32.Foundation;
-using Windows.Win32.UI.WindowsAndMessaging;
 
 namespace DevHome.PI.ViewModels;
 
@@ -42,6 +38,9 @@ public partial class BarWindowViewModel : ObservableObject
 
     private readonly ObservableCollection<Button> _externalTools = [];
     private readonly SnapHelper _snapHelper;
+
+    [ObservableProperty]
+    private PIInsightsService _insightsService;
 
     [ObservableProperty]
     private string _systemCpuUsage = string.Empty;
@@ -145,6 +144,9 @@ public partial class BarWindowViewModel : ObservableObject
 
         ((INotifyCollectionChanged)ExternalToolsHelper.Instance.FilteredExternalTools).CollectionChanged += FilteredExternalTools_CollectionChanged;
         FilteredExternalTools_CollectionChanged(null, null);
+
+        _insightsService = Application.Current.GetService<PIInsightsService>();
+        _insightsService.PropertyChanged += InsightsService_PropertyChanged;
     }
 
     partial void OnShowingExpandedContentChanged(bool value)
@@ -349,6 +351,15 @@ public partial class BarWindowViewModel : ObservableObject
         }
     }
 
+    private void InsightsService_PropertyChanged(object? sender, PropertyChangedEventArgs e)
+    {
+        if (e.PropertyName == nameof(PIInsightsService.UnreadCount))
+        {
+            UnreadInsightsCount = InsightsService.UnreadCount;
+            InsightsBadgeOpacity = UnreadInsightsCount > 0 ? 1 : 0;
+        }
+    }
+
     public void ManageExternalToolsButton_ExternalToolLaunchRequest(object sender, ExternalTool tool)
     {
         tool.InvokeTool(null, TargetAppData.Instance.TargetProcess?.Id, TargetAppData.Instance.HWnd);
@@ -358,12 +369,6 @@ public partial class BarWindowViewModel : ObservableObject
     public void LaunchAdvancedAppsPageInWindowsSettings()
     {
         _ = Launcher.LaunchUriAsync(new("ms-settings:advanced-apps"));
-    }
-
-    public void UpdateUnreadInsightsCount(int count)
-    {
-        UnreadInsightsCount = count;
-        InsightsBadgeOpacity = count > 0 ? 1 : 0;
     }
 
     [RelayCommand]
