+++ tools/PI/DevHome.PI/Views/BarWindow.cs
@@ -176,14 +176,4 @@ public partial class BarWindow
     {
         return _viewModel.IsSnapped;
     }
-
-    public void IncreaseUnreadInsightsCount()
-    {
-        _viewModel.UpdateUnreadInsightsCount(_viewModel.UnreadInsightsCount + 1);
-    }
-
-    public void UpdateUnreadInsightsCount(int count)
-    {
-        _viewModel.UpdateUnreadInsightsCount(count);
-    }
 }
